refactor(bbox): moved bbox calculators into utils folder
These calculators can be used by every chart type, so it's was moved from the specific lib/axes
folder
