+++ src/utils/bbox/bbox_calculator.ts
