+++ src/utils/bbox/canvas_text_bbox_calculator.test.ts
