+++ src/utils/bbox/svg_text_bbox_calculator.ts
